Rename the module sogo3 -> sogo
On Debian Buster, the sogo package is in fact SOGo 4.  FreeBSD has
www/sogo3 and www/sogo4 to choose from, but both install a similar
configuration to the same directory (SOGo 4 is a continuation of SOGo 3,
and the update is straighforward).

Rename the package so that the user does not expect it to install SOGo 3
by default, and actually install SOGo 4 when both SOGo 3 and 4 are
available.